Got rid of indexed files that should be ignored. Some minor updates that does not really do anything.

+++ EdgeDetection/EdgeDetection/EdgeDetection/hough.h
@@ -45,6 +45,6 @@ public:
 houghSpace* houghTransform(edgeImage* edgeimage, int angleResolution, float angleMargin);
 houghSpace* houghFiltering(houghSpace *hough, int threshold, int windowWidth, int windowHeight);
 grayImage* houghRecronstruction(int width, int height, houghSpace *hough);
-list<quadrangle*>* houghIdentifyQuadrangles(houghSpace *hough, edgeImage *edgeImage, int horizontalDistance, double horizontalAngMargin, int perpendicularDistance, double perpendicularAngMargin, double holeTolerance, int continousHoleTolerance);
+list<quadrangle*>* houghIdentifyQuadrangles(houghSpace *hough, edgeImage *edgeImage, int horizontalDistance, float horizontalAngMargin, int perpendicularDistance, float perpendicularAngMargin, double holeTolerance, int continousHoleTolerance);
 
 #endif
